+++ tsconfig.json
@@ -6,6 +6,7 @@
     "module": "commonjs",
     "outDir": "./dist",
     "strict": true,
+    "declaration": true,
     "esModuleInterop": true,
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true
